Host_A/requirements.txt
kagglehub
kafka-python
